% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irrmesh}
\alias{irrmesh}
\title{Irrlicht 3D meshplot}
\usage{
irrmesh(pathmesh, pathtexture, driverselect)
}
\arguments{
\item{pathmesh}{path to mesh (see supported formats here: \url{http://irrlicht.sourceforge.net/?page_id=45})}

\item{pathtexture}{path to texture (see supported formats here: \url{http://irrlicht.sourceforge.net/?page_id=45})}

\item{driverselect}{"a": OPENGL,
"b": DIRECT3D9
"c": DIRECT3D8
"d": BURNINGSVIDEO
"e": SOFTWARE
"f": NULL}
}
\value{
boolean value - side effect irrlicht window is relevant
}
\description{
Creates a 3D meshplot using the Irrlicht engine.
}
\details{
For test data see for example \url{https://github.com/inexor-game/data-playground/tree/master/model/map/modelpack2}.
}
\examples{
#irrmesh(
# "bspmesh.md2", 
# "bsptexture.bmp", 
# "a"
#)

}

