% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irrscatter}
\alias{irrscatter}
\title{test}
\usage{
irrscatter(x, y, z, size, driverselect)
}
\arguments{
\item{driverselect}{}
}
\value{
test
}
\description{
test
}
\examples{
irrscatter(x = rnorm(500)*200, y = rnorm(500)*200, z = rnorm(500)*200, size = rnorm(500)*10, driverselect = "a")

}

