% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh2ply_string.R
\name{mesh2ply_string}
\alias{mesh2ply_string}
\title{mesh2ply string version}
\usage{
mesh2ply_string(mesh)
}
\arguments{
\item{mesh}{truet}
}
\value{
string that describes a mesh in ply format
}
\description{
Mask for Morpho::mesh2ply
}
