% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plot_irr}
\alias{plot_irr}
\title{Irrlicht 3D plot}
\usage{
plot_irr(points_df = NULL, mesh_string_input = NULL, doomhud = FALSE,
  video_driver = "a")
}
\arguments{
\item{points_df}{optional - data.frame with coordinates and attributes 
of scattered points}

\item{mesh_string_input}{floet}

\item{doomhud}{fluet}

\item{video_driver}{\itemize{
  \item{"a": }{OPENGL (default)} 
  \item{"b": }{DIRECT3D9}
  \item{"c": }{DIRECT3D8}
  \item{"d": }{BURNINGSVIDEO}
  \item{"e": }{SOFTWARE}
  \item{"f": }{NULL}
}}
}
\value{
boolean value - side effect; irrlicht window is relevant
}
\description{
Creates a 3D plot using the Irrlicht engine.
}
\examples{
\dontrun{
plot_irr(
  points = data.frame(
    x = rnorm(500)*200, 
    y = rnorm(500)*200, 
    z = rnorm(500)*200,
    size = rnorm(500)*10 
  )
)
}

}
