% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plot_irr}
\alias{plot_irr}
\title{Irrlicht 3D scatterplot}
\usage{
plot_irr(x, y, z, size, driverselect)
}
\arguments{
\item{x}{vector with x-axis coordinates}

\item{y}{vector with y-axis coordinates}

\item{z}{vector with z-axis coordinates}

\item{size}{vector with size values}

\item{driverselect}{"a": OPENGL,
"b": DIRECT3D9
"c": DIRECT3D8
"d": BURNINGSVIDEO
"e": SOFTWARE
"f": NULL}
}
\value{
boolean value - side effect irrlicht window is relevant
}
\description{
Creates a 3D scatterplot using the Irrlicht engine.
}
\examples{
#plot_irr(
#  x = rnorm(500)*200, 
#  y = rnorm(500)*200, 
#  z = rnorm(500)*200, 
#  size = rnorm(500)*10, 
#  driverselect = "a"
#)

}
